requirements.txt
boto3==1.35.39
grpcio==1.66.2
protobuf==5.28.2
redis==5.1.1
